* Object_emC: new operation name_ClassJc and name_FieldJc returns a StringJc.   This is the only one proper Method to get Information because the name field is not 0-terminated.   It works proper with StringBuilderJcpp::operator+(StringJc) or append_s_StringBuilderJc. * Object_emC creates systemInit_emC. * Exception_emC: new operation writeException(...) produce one line with relevant information, used for Simulink message. * stdef_SmlkAccelerator/applstdef_emC.h yet with __DONOTUSE_REFLECTION__, may be temporarry TODO * Handle_ptr64_emC.c: now works without ShardedMem, especially in an win64-appl with uint32 handle.   Usage of ShardedMem is only switched on with __SIMULINK_SFN__. Only there a global common area is need   (same situation with dll, not regarded now). * os_mem: os_freeMem(null) it admissible. * sourceSpecials/osal_Windows/os_time was twice for Windows, only one version now in osal_Windows_Msc15 * bugfixes compiling StringBuilder usage in FormatterJc. Building the whole library. * J1c/*.c: Some adaption to throw_(...__FILE__)

+++ source/Jc/FileIoJc.c
@@ -169,7 +169,7 @@ void write_FileOutputStreamJc  (  FileOutputStreamJc_s* ythis, void* data, int o
   STACKTRC_TENTRY("write_FileOutputStreamJc");
   nrofBytesWritten = os_fwrite(ythis->file, addOffset_MemUnit(data, offset), len);
   if(nrofBytesWritten != len){
-    THROW1_s0(IOException, "writing file", (int)ythis);
+    THROW1_s0(IOException, "writing file", (int)(intptr_t)ythis);
 
   }
   STACKTRC_LEAVE;
@@ -259,7 +259,7 @@ void write_FileWriterJc  (  FileWriterJc_s* ythis, StringJc text, ThCxt* _thCxt)
   STACKTRC_TENTRY("write_FileWriterJc");
   sText = getCharsAndLength_StringJc(&text, &nrofChars);
   error = os_fwrite(ythis->file, sText, nrofChars);
-  if(error < 0){ THROW1_s0(IOException, "writing file", (int)ythis);}
+  if(error < 0){ THROW1_s0(IOException, "writing file", (int)(intptr_t)ythis);}
   STACKTRC_LEAVE; 
 }
