+++ source/Jc/ReflectionJc.c
@@ -159,10 +159,11 @@ FieldJc_Y* new_FieldJc_Y(int size)
   return thiz;
 }
 
-
+//Note: use name_ClassJc
 const char* getName_ClassJc(ClassJc const* thiz)
 { if( ((uint64)(thiz)) < kREFLECTION_LastConstant )
   { thiz = simpleTypes_ClassJc[(int)(thiz)];
+    //Note, TODO thiz should not be a simpleType-const
   }
   return thiz->name;
 }
@@ -180,8 +181,9 @@ METHOD_C bool isPrimitive_ClassJc(ClassJc const* thiz)
 
 METHOD_C int32 getModifiers_ClassJc(ClassJc const* thiz)
 {
-  if( ((uint64)(thiz)) < kREFLECTION_LastConstant )
+  if( ((intptr_t)(thiz)) < kREFLECTION_LastConstant )
   { thiz = simpleTypes_ClassJc[(int)(thiz)];
+    //Note, TODO thiz should not be a simpleType-const
   }
   return thiz->modifiers;
 }
@@ -199,8 +201,9 @@ METHOD_C ClassJc const* getEnclosingClass_ClassJc(ClassJc const* thiz)
  */
 FieldJc_Y const* getDeclaredFields_ClassJc(const ClassJc* thiz)
 { FieldJc_Y const* fields;
-  if( ((uint64)(thiz)) < kREFLECTION_LastConstant ) //2015-06 regard 64-bit-addresses and >0x7fffffff
+  if( ((intptr_t)(thiz)) < kREFLECTION_LastConstant ) //2015-06 regard 64-bit-addresses and >0x7fffffff
   { thiz = simpleTypes_ClassJc[(int)(thiz)];
+    //Note, TODO thiz should not be a simpleType-const
   }
   //thiz is the real class:
   fields = thiz->attributes;
